requirements.txt
langchain-openai==0.1.8
langchain-chroma==0.1.1
langchain==0.2.5
langchain-community==0.2.5
beautifulsoup4==4.12.3
pypdf==4.2.0
pdfminer
sentence_transformers==3.0.1
langchain_huggingface==0.0.3